+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 0.1.3 (2014-01-20)
+
+* Added `QueueHandler`, inherits standard `logging.Handler` that `emit` to any standard `Queue` compatible implementations, including all the `Queue` implementations in this library
+* Exposed accessibility of all classes (`Queue` implementations and `logging.Handler` implementations to bypass module namespace)
+* Behavior of `CounterQueue.get` has changed from just returning the item value to a `tuple` of item value and its number of appearances in the queue, eg: `('dota', 2)`; an empty `tuple` is returned when the `CounterQueue` is empty
+
 ### 0.1.2 (2014-01-17)
 
 * Added `CountQueue`, inherits `UniQueue` but the score is used as a count of item appearance, that the item has the highest count gets placed in front to be `get` first